requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
passlib==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.0